move getJwtToken into the HTTP request handler
